Templates and Menu
Now instead of having to edit the menu items in each .html file you can
just edit the _menu.html file and the new menu will be loaded by all the
pages.

_template.html is a HTML static type template that currently causes
error.
WE ARE GOING TO NEED TO CREATE A .JS SCRIPT FOR EACH PAGE LIKE WE HAVE
FOR THE PROFILE AND GAMING TOOLS.